Rename: RedisEntity 패키지 변경
